feat: forcing reset wallet when it was already loaded before for the testnet
